third_party/WebKit/LayoutTests/fast/xpath/evaluator-exceptions.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="console"></div>
<script>
    var evalulator = new XPathEvaluator;
    var nsResolver = evalulator.createNSResolver(document);

    shouldThrow("evalulator.evaluate('/body', 0, nsResolver, 0, null);");

    var fragment = document.createDocumentFragment(); 
    shouldThrow("evalulator.evaluate('/body', fragment, nsResolver, 0, null);");

    var doctype = document.doctype; 
    shouldThrow("evalulator.evaluate('/body', doctype, nsResolver, 0, null);");
</script>
</body>
</html>
